requirements.txt
torch
torchvision
timm
streamlit
albumentations
fastapi
uvicorn
wandb
numpy
pandas
pillow
scikit-learn
